(dbg) got lost, refactored a bit, this is probably going to fails badly :'(?
